requirements.txt
Flask==2.3.3
requests==2.31.0
gspread==5.11.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
qrcode[pil]==7.4.2
python-dotenv==1.0.0
gunicorn==21.2.0